Multer and modification post, put, delete book

+++ controllers/Book.js
@@ -1,18 +1,15 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.addBook = (req, res, next) => {
-    const token = req.headers.authorization;
-    if (!token) {
-        return res.status(401).json({ message: 'Token manquant dans les en-têtes de la requête' });
-    }
-    console.log(token)
-    console.log("Données reçues pour le livre :", req.body);
-
-    delete req.body._id
+    const bookObject = JSON.parse(req.body.book);
+    delete bookObject._id;
+    delete bookObject._userId;
     const book = new Book({
-        ...req.body
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
-    console.log("Objet Book créé :", book);
 
     book.save()
         .then(() => res.status(201).json({ message: 'Livre enregistré'}))
@@ -20,15 +17,49 @@ exports.addBook = (req, res, next) => {
   };
 
 exports.modifyBook =  (req, res, next) => {
-    Book.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-        .then(() => res.status(200).json({message: 'Livre modifié'}))
-        .catch(error => res.status(400).json({error}));
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        } : { ...req.body };
+
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(403).json({ message : '403: unauthorized request' });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                req.file && fs.unlink(`images/${filename}`, (err => {
+                        if (err) console.log(err);
+                    })
+                );
+                Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Livre modifié' }))
+                    .catch(error => res.status(400).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(404).json({ error });
+        });
     };
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({_id: req.params.id})
-        .then(() => res.status(200).json({message: 'Livre supprimé'}))
-        .catch(error => res.status(400).json({error}));
+    Book.findOne({ _id: req.params.id })
+    .then(book => {
+        if (book.userId != req.auth.userId) {
+            res.status(403).json({ message: '403: unauthorized request' });
+        } else {
+            const filename = book.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Book.deleteOne({ _id: req.params.id })
+                    .then(() => { res.status(200).json({ message: 'Objet supprimé !' }) })
+                    .catch(error => res.status(400).json({ error }));
+            });
+        }
+    })
+    .catch( error => {
+        res.status(404).json({ error });
+    });
     };
 
 exports.getAllBooks = (req, res, next) => {
